requirements.txt
fastapi
uvicorn
pandas
scikit-learn
langchain
langchain-community
langchain-openai
faiss-cpu
openai
tiktoken
python-dotenv
python-docx
openpyxl